use typeshare quotes
